arreglos en database
