+++ api/src/connection/db.js
@@ -15,9 +15,11 @@ const {
 /******* FIN DE EXPORTACIONES********** */
 
 const sequelize = new Sequelize(`postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}`, {
-  logging: false, // set to console.log to see the raw SQL queries
-  native: false, // lets Sequelize know we can use pg-native for ~30% more speed
+logging: false, // set to console.log to see the raw SQL queries
+native: false, // lets Sequelize know we can use pg-native for ~30% more speed
 });
+
+
 
 //creacion de modelos
 userModel(sequelize)
